Change language component to language-switcher component
